+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/dto/ChallengeDto.java
@@ -22,7 +22,7 @@ public class ChallengeDto {
     private UUID challengeId;
 
     @JsonProperty(value = "challenge_title", index = 1)
-    private Map<Locale, String> title;
+    private String title;
 
     @JsonProperty(index = 2)
     private String level;