+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/repository/ChallengeRepository.java
@@ -32,9 +32,10 @@ public interface ChallengeRepository extends ReactiveSortingRepository<Challenge
     @Query(value = "{ 'languages.idLanguage' : ?0 }")
     Flux<ChallengeDocument> findByLanguages_IdLanguage(UUID idLanguage);
     Flux<ChallengeDocument> findByLanguages_LanguageName(String languageName);
+    @Query(value = "{ 'challenge_title' : { $regex: ?0, $options: 'i' } }", exists = true)
+    Mono<Boolean> existsByChallengeTitle(String title);
     @Query(value = "{ 'challenge_title.ca' : { $regex: ?0, $options: 'i' } }", exists = true)
     Mono<Boolean> existsByChallengeTitleCa(String title);
     Flux<ChallengeDocument> findByTopic(Topic topic);
 
 }
-